pyproject.toml
[tool.poetry]
name = "diffagent"
version = "1.0.0"
description = "AI-powered CLI tool that validates configuration file changes to prevent misconfigurations"
authors = ["Sebastian Simon <Bastisimon95@googlemail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/your-org/DiffAgent"
keywords = ["configuration", "validation", "git", "pre-commit", "llm", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Version Control :: Git",
]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.3.0"
langchain-core = "^0.3.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
openai = "^1.0.0"
anthropic = "^0.40.0"
cfgnet = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.0.0"
ruff = "^0.3.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
diffagent = "cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=. --cov-report=term-missing"
